Clean up RegexNodeToSymbolicConverter
